packages/core/pyproject.toml
[project]
name = "rag-core"
version = "0.1.0"
description = "Shared core library for RAG agent platform"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-cohere>=0.4.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "qdrant-client>=1.12.0",
    "fastembed>=0.4.0",
    "google-genai>=1.0.0",
    "langsmith>=0.3.0",
    "asyncpg>=0.29.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "rapidfuzz>=3.0.0",
    "langchain-qdrant>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core"]
